refactor excel_exporter module and improve docs
