+++ excel_exporter/__init__.py
@@ -1,0 +1,9 @@
+"""
+Excel Exporter Package
+
+Exports JIRA metrics to formatted Excel workbooks with charts and table data.
+"""
+
+from .exporter import ExcelExporter
+
+__all__ = ['ExcelExporter']
